refactor(llgo/rust/hyper): Refactor the location of the code
