+++ go.mod
@@ -2,4 +2,4 @@ module github.com/goplus/llgoexamples
 
 go 1.20
 
-require github.com/goplus/llgo v0.9.1-0.20240712060421-858d38d314a3
+require github.com/goplus/llgo v0.9.3-0.20240726020431-98d075728f2b
