+++ rust/hyper/hyper.go
@@ -1,15 +1,9 @@
 package hyper
 
 import (
-	"fmt"
-	"strings"
 	_ "unsafe"
 
 	"github.com/goplus/llgo/c"
-	"github.com/goplus/llgo/c/net"
-	"github.com/goplus/llgo/c/os"
-	"github.com/goplus/llgo/c/sys"
-	"github.com/goplus/llgo/c/syscall"
 )
 
 const (
@@ -450,387 +444,3 @@ func (waker *Waker) Free() {}
 // Wake up the task associated with a waker.
 // llgo:link (*Waker).Wake C.hyper_waker_wake
 func (waker *Waker) Wake() {}
-
-type ConnData struct {
-	Fd         c.Int
-	ReadWaker  *Waker
-	WriteWaker *Waker
-}
-
-func ConnectTo(host string, port string) c.Int {
-	c.Printf(c.Str("connecting to port %s on %s...\n"), c.AllocaCStr(port), c.AllocaCStr(host))
-	var hints net.AddrInfo
-	hints.Family = net.AF_UNSPEC
-	hints.SockType = net.SOCK_STREAM
-
-	var result, rp *net.AddrInfo
-
-	if net.Getaddrinfo(c.AllocaCStr(host), c.AllocaCStr(port), &hints, &result) != 0 {
-		panic(fmt.Sprintf("dns failed for %s\n", host))
-	}
-
-	var sfd c.Int
-	for rp = result; rp != nil; rp = rp.Next {
-		sfd = net.Socket(rp.Family, rp.SockType, rp.Protocol)
-		if sfd == -1 {
-			continue
-		}
-		if net.Connect(sfd, rp.Addr, rp.AddrLen) != -1 {
-			break
-		}
-		os.Close(sfd)
-	}
-
-	net.Freeaddrinfo(result)
-
-	// no address succeeded
-	if rp == nil || sfd < 0 {
-		panic(fmt.Sprintf("connect failed for %s\n", host))
-	}
-
-	c.Printf(c.Str("connected to %s\n"), c.AllocaCStr(host))
-
-	if os.Fcntl(sfd, os.F_SETFL, os.O_NONBLOCK) != 0 {
-		panic("failed to set net to non-blocking\n")
-	}
-	return sfd
-}
-
-func ReadCallBack(userdata c.Pointer, ctx *Context, buf *uint8, bufLen uintptr) uintptr {
-	conn := (*ConnData)(userdata)
-
-	ret := os.Read(conn.Fd, c.Pointer(buf), bufLen)
-
-	if ret >= 0 {
-		return uintptr(ret)
-	}
-
-	if os.Errno != os.EAGAIN {
-		c.Perror(c.Str("[read callback fail]"))
-		// kaboom
-		return IoError
-	}
-
-	// would block, register interest
-	if conn.ReadWaker != nil {
-		conn.ReadWaker.Free()
-	}
-	conn.ReadWaker = ctx.Waker()
-	return IoPending
-}
-
-func WriteCallBack(userdata c.Pointer, ctx *Context, buf *uint8, bufLen uintptr) uintptr {
-	conn := (*ConnData)(userdata)
-	ret := os.Write(conn.Fd, c.Pointer(buf), bufLen)
-
-	if int(ret) >= 0 {
-		return uintptr(ret)
-	}
-
-	if os.Errno != os.EAGAIN {
-		c.Perror(c.Str("[write callback fail]"))
-		// kaboom
-		return IoError
-	}
-
-	// would block, register interest
-	if conn.WriteWaker != nil {
-		conn.WriteWaker.Free()
-	}
-	conn.WriteWaker = ctx.Waker()
-	return IoPending
-}
-
-func PrintEachHeader(userdata c.Pointer, name *uint8, nameLen uintptr, value *uint8, valueLen uintptr) c.Int {
-	//c.Printf(c.Str("%.*s: %.*s\n"), int(nameLen), name, int(valueLen), value)
-	return IterContinue
-}
-
-func PrintEachChunk(userdata c.Pointer, chunk *Buf) c.Int {
-	buf := chunk.Bytes()
-	len := chunk.Len()
-
-	os.Write(1, c.Pointer(buf), len)
-
-	return IterContinue
-}
-
-func FreeConnData(conn *ConnData) {
-	if conn.ReadWaker != nil {
-		conn.ReadWaker.Free()
-		conn.ReadWaker = nil
-	}
-	if conn.WriteWaker != nil {
-		conn.WriteWaker.Free()
-		conn.WriteWaker = nil
-	}
-}
-
-func NewConnData(fd c.Int) *ConnData {
-	return &ConnData{Fd: fd, ReadWaker: nil, WriteWaker: nil}
-}
-
-func NewIoWithConnReadWrite(connData *ConnData) *Io {
-	io := NewIo()
-	io.SetUserdata(c.Pointer(connData))
-	io.SetRead(ReadCallBack)
-	io.SetWrite(WriteCallBack)
-	return io
-}
-
-func (task *Task) SetUserData(userData ExampleId) {
-	var data = userData
-	task.SetUserdata(c.Pointer(uintptr(data)))
-}
-
-type RequestConfig struct {
-	ReqMethod      string
-	ReqHost        string
-	ReqPort        string
-	ReqUri         string
-	ReqHeaders     map[string]string
-	ReqHTTPVersion HTTPVersion
-	TimeoutSec     int64
-	TimeoutUsec    int32
-	//ReqBody
-	//ReqURIParts
-}
-
-type RequestResponse struct {
-	Status  uint16
-	Message string
-	Content []byte
-}
-
-func SendRequest(requestConfig *RequestConfig) *RequestResponse {
-	// Parsing request configuration
-	method := strings.ToUpper(requestConfig.ReqMethod)
-	if method == "" {
-		method = "GET"
-	}
-	if method != "GET" && method != "POST" {
-		panic(fmt.Sprintf("There is no method called %s\n", method))
-	}
-	host := requestConfig.ReqHost
-	port := requestConfig.ReqPort
-	path := requestConfig.ReqUri
-	headers := requestConfig.ReqHeaders
-
-	req := NewRequest()
-
-	// Prepare the request
-	// Set the request method and uri
-	if req.SetMethod((*uint8)(&[]byte(method)[0]), c.Strlen(c.AllocaCStr(method))) != OK {
-		panic(fmt.Sprintf("error setting method %s\n", method))
-	}
-	if req.SetURI((*uint8)(&[]byte(path)[0]), c.Strlen(c.AllocaCStr(path))) != OK {
-		panic(fmt.Sprintf("error setting uri %s\n", path))
-	}
-
-	// Set the request headers
-	reqHeaders := req.Headers()
-	if reqHeaders.Set((*uint8)(&[]byte("Host")[0]), c.Strlen(c.Str("Host")), (*uint8)(&[]byte(host)[0]), c.Strlen(c.AllocaCStr(host))) != OK {
-		panic("error setting headers\n")
-	}
-
-	if headers != nil && len(headers) > 0 {
-		for k, v := range headers {
-			if reqHeaders.Set((*uint8)(&[]byte(k)[0]), c.Strlen(c.AllocaCStr(k)), (*uint8)(&[]byte(v)[0]), c.Strlen(c.AllocaCStr(v))) != OK {
-				panic("error setting headers\n")
-			}
-		}
-	}
-
-	// Set the request http version
-	if requestConfig.ReqHTTPVersion != 0 {
-		req.SetVersion(c.Int(requestConfig.ReqHTTPVersion))
-	}
-
-	// TODO set the uri parts
-	// TODO set the request body
-
-	//var response RequestResponse
-
-	fd := ConnectTo(host, port)
-
-	connData := NewConnData(fd)
-
-	// Hookup the IO
-	io := NewIoWithConnReadWrite(connData)
-
-	// We need an executor generally to poll futures
-	exec := NewExecutor()
-
-	// Prepare client options
-	opts := NewClientConnOptions()
-	opts.Exec(exec)
-
-	handshakeTask := Handshake(io, opts)
-	handshakeTask.SetUserData(ExampleHandshake)
-
-	// Let's wait for the handshake to finish...
-	exec.Push(handshakeTask)
-
-	var fdsRead, fdsWrite, fdsExcep syscall.FdSet
-	var err *Error
-	var response RequestResponse
-
-	// The polling state machine!
-	for {
-		// Poll all ready tasks and act on them...
-		for {
-			task := exec.Poll()
-
-			if task == nil {
-				break
-			}
-
-			switch (ExampleId)(uintptr(task.Userdata())) {
-			case ExampleHandshake:
-				if task.Type() == TaskError {
-					c.Printf(c.Str("handshake error!\n"))
-					err = (*Error)(task.Value())
-					fail(err)
-				}
-				if task.Type() != TaskClientConn {
-					c.Printf(c.Str("169 unexpected task type\n"))
-					fail(err)
-				}
-
-				c.Printf(c.Str("preparing http request ...\n"))
-
-				client := (*ClientConn)(task.Value())
-				task.Free()
-
-				// Send it!
-				sendTask := client.Send(req)
-				sendTask.SetUserData(ExampleSend)
-				c.Printf(c.Str("sending ...\n"))
-				sendRes := exec.Push(sendTask)
-				if sendRes != OK {
-					panic("error send\n")
-				}
-
-				// For this example, no longer need the client
-				client.Free()
-
-				break
-			case ExampleSend:
-				if task.Type() == TaskError {
-					c.Printf(c.Str("send error!\n"))
-					err = (*Error)(task.Value())
-					fail(err)
-				}
-				if task.Type() != TaskResponse {
-					c.Printf(c.Str("unexpected task type\n"))
-					fail(err)
-				}
-
-				// Take the results
-				resp := (*Response)(task.Value())
-				task.Free()
-
-				rp := resp.ReasonPhrase()
-				rpLen := resp.ReasonPhraseLen()
-
-				response.Status = resp.Status()
-				response.Message = string((*[1 << 30]byte)(c.Pointer(rp))[:rpLen:rpLen])
-
-				headers := resp.Headers()
-				headers.Foreach(PrintEachHeader, nil)
-				c.Printf(c.Str("\n"))
-
-				respBody := resp.Body()
-				foreachTask := respBody.Foreach(PrintEachChunk, nil)
-				foreachTask.SetUserData(ExampleRespBody)
-				exec.Push(foreachTask)
-
-				// No longer need the response
-				resp.Free()
-
-				break
-			case ExampleRespBody:
-				if task.Type() == TaskError {
-					c.Printf(c.Str("body error!\n"))
-					err = (*Error)(task.Value())
-					fail(err)
-				}
-				if task.Type() != TaskEmpty {
-					c.Printf(c.Str("unexpected task type\n"))
-					fail(err)
-				}
-
-				// Cleaning up before exiting
-				task.Free()
-				exec.Free()
-				FreeConnData(connData)
-
-				return &response
-			case ExampleNotSet:
-				// A background task for hyper_client completed...
-				task.Free()
-				break
-			}
-		}
-
-		// All futures are pending on IO work, so select on the fds.
-
-		sys.FD_ZERO(&fdsRead)
-		sys.FD_ZERO(&fdsWrite)
-		sys.FD_ZERO(&fdsExcep)
-
-		if connData.ReadWaker != nil {
-			sys.FD_SET(connData.Fd, &fdsRead)
-		}
-		if connData.WriteWaker != nil {
-			sys.FD_SET(connData.Fd, &fdsWrite)
-		}
-
-		// Set the request timeout
-		var tv syscall.Timeval
-		if requestConfig.TimeoutSec != 0 || requestConfig.TimeoutUsec != 0 {
-			tv.Sec = requestConfig.TimeoutSec
-			tv.Usec = requestConfig.TimeoutUsec
-		} else {
-			tv.Sec = 10
-		}
-
-		selRet := sys.Select(connData.Fd+1, &fdsRead, &fdsWrite, &fdsExcep, &tv)
-		if selRet < 0 {
-			panic("select() error\n")
-		} else if selRet == 0 {
-			panic("select() timeout\n")
-		}
-
-		if sys.FD_ISSET(connData.Fd, &fdsRead) != 0 {
-			connData.ReadWaker.Wake()
-			connData.ReadWaker = nil
-		}
-
-		if sys.FD_ISSET(connData.Fd, &fdsWrite) != 0 {
-			connData.WriteWaker.Wake()
-			connData.WriteWaker = nil
-		}
-	}
-}
-
-func fail(err *Error) {
-	if err != nil {
-		c.Printf(c.Str("error code: %d\n"), err.Code())
-		// grab the error details
-		var errBuf [256]c.Char
-		errLen := err.Print((*uint8)(c.Pointer(&errBuf[:][0])), uintptr(len(errBuf)))
-
-		//c.Printf(c.Str("details: %.*s\n"), c.int(errLen), errBuf)
-		c.Printf(c.Str("details: "))
-		for i := 0; i < int(errLen); i++ {
-			c.Printf(c.Str("%c"), errBuf[i])
-		}
-		c.Printf(c.Str("\n"))
-
-		// clean up the error
-		err.Free()
-		panic("request failed\n")
-	}
-	return
-}
